+++ src/fs.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include<iostream>
-#include<stdio.h>
+#include <string>
 
 constexpr int FAILED = -1;
 
@@ -9,10 +8,10 @@ constexpr int FAILED = -1;
 //
 // 读取文件成功返回缓冲区长度, 失败返回 -1
 //
-int readFile(std::string& file, char **buf);
+size_t readFile(std::string& file, char **buf);
 
 
 //
 // 读取文本文件(utf8)成功返回缓冲区长度, 失败返回 -1
 //
-int readTxtFile(std::string& fileName, std::string& content);
+size_t readTxtFile(std::string& fileName, std::string& content);
